Get rid of roles
