+++ lib/git_host.rb
