+++ lib/static_model.rb
