+++ lib/votes.rb
