+++ spec/lib/issue_commonality_spec.rb
