+++ spec/lib/votes_spec.rb
