+++ spec/models/user_spec.rb
@@ -185,4 +185,14 @@ describe User do
 
     it { User.not_in_project(@project).should == [@user, @project.owner] }
   end
+
+  describe 'normal user' do
+    let(:user) { create(:user, name: 'John Smith') }
+
+    it { user.is_admin?.should be_false }
+    it { user.require_ssh_key?.should be_true }
+    it { user.can_create_group?.should be_false }
+    it { user.can_create_project?.should be_true }
+    it { user.first_name.should == 'John' }
+  end
 end
